Fix flaky custom data tests and move into their own class
